+++ lostpixel.config.ts
@@ -1,0 +1,9 @@
+import { CustomProjectConfig } from 'lost-pixel';
+
+export const config: CustomProjectConfig = {
+  storybookShots: {
+    storybookUrl: './storybook-static',
+  },
+  lostPixelProjectId: 'cm4lbix892lch142enxr02fpo',
+  apiKey: process.env.LOST_PIXEL_API_KEY,
+};