+++ src/Style/index.ts
@@ -6,3 +6,4 @@ export * from './PillComponentStyle';
 export * from './ShadowStyle';
 export * from './FontStyle';
 export * from './InlineTextComponentStyle';
+export * from './TmpStyle';