+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.0.12 - 2024-12-16
 ### Added
 - Add support for i18n.
-- Integrate Storybook and visual regression tests, closes #52.
+- Integrate Storybook and visual regression tests, closes [#52].
+
 ### Changed
 - Improve documentation.
+
 ### Fixed
 - Fix serif font family.